mythavtest: rename commandlineparser.* to mythavtest_...
